+++ src/media/ImageCache.h
@@ -1,27 +1,29 @@
 #pragma once
 
+#include "media/AsyncImage.h"
 #include "media/Path.h"
 
-#include <QHash>
-#include <QImage>
 #include <QSize>
 #include <QString>
-#include <QVector>
 
 class ImageCache : public QObject
 {
     Q_OBJECT
 public:
     explicit ImageCache(QObject* parent = nullptr);
-    static ImageCache* instance(QObject* parent = nullptr);
-    QImage image(mediaelch::FilePath path, int width, int height, int& origWidth, int& origHeight);
-    QSize imageSize(mediaelch::FilePath path);
-    void invalidateImages(mediaelch::FilePath path);
+    static ImageCache* instance();
+
+    void invalidateImages(const mediaelch::FilePath& path);
     void clearCache();
+
+    /// \brief Asynchronously load the given image and resize it. Caches the resized result on disk.
+    /// \details
+    ///   If either width or height of the targetSize is 0, the aspect ratio is respected and the
+    ///   not-null value used. See mediaelch::scaledImage() for details.  The resized result
+    ///   is cached on disk and loaded instead of the original on subsequent requests.
+    ///   No in-memory caching is performed.
+    std::unique_ptr<mediaelch::AsyncImage> loadImageAsync(const mediaelch::FilePath& path, QSize targetSize);
 
 private:
     mediaelch::DirectoryPath m_cacheDir;
-    QHash<mediaelch::FilePath, QVector<qint64>> m_lastModifiedTimes;
-    QImage scaledImage(QImage img, int width, int height);
-    qint64 getLastModified(const mediaelch::FilePath& fileName);
 };