+++ src/ui/tv_show/TvShowWidgetSeason.cpp
@@ -159,8 +159,8 @@ void TvShowWidgetSeason::updateImages(QSet<ImageType> images)
                         .isEmpty()
                    && (!m_show->imagesToRemove().contains(imageType)
                        || !m_show->imagesToRemove().value(imageType).contains(m_season))) {
-            image->setImage(
-                Manager::instance()->mediaCenterInterfaceTvShow()->imageFileName(m_show, imageType, m_season));
+            image->setImageFromPath(mediaelch::FilePath{
+                Manager::instance()->mediaCenterInterfaceTvShow()->imageFileName(m_show, imageType, m_season)});
         }
     }
 }